Rename env var to AUTHORIZATION_MANAGEMENT_URL
