+++ packages/authorization-updater/.env
@@ -3,14 +3,14 @@ PORT=3000
 LOG_LEVEL=info
 
 KAFKA_CLIENT_ID="authorization-updater-client"
-KAFKA_GROUP_ID="authorization-updater-group"
+KAFKA_GROUP_ID="authorization-updater-group-local"
 KAFKA_BROKERS="localhost:9092"
 KAFKA_DISABLE_AWS_IAM_AUTH="true"
 KAFKA_TOPICS="event-store.catalog.events"
 AWS_REGION="eu-central-1"
 
 SKIP_JWT_VERIFICATION=true
-AUTH_MNGT_SERVICE_URL="/authorization-management/0.0"
+AUTHORIZATION_MANAGEMENT_URL="/authorization-management/0.0"
 RSA_KEYS_IDENTIFIERS=17c1177f-d7dc-4181-9f54-6fd416bf229b
 KMS_MAX_ACQUISITION_TIMEOUT_SECONDS=60
 GENERATED_JWT_SECONDS_TO_EXPIRE=60
