+++ packages/commons/src/config/authServerConfig.ts
@@ -2,10 +2,10 @@ import { z } from "zod";
 
 export const AuthServiceConfig = z
   .object({
-    AUTH_MNGT_SERVICE_URL: z.string(),
+    AUTHORIZATION_MANAGEMENT_URL: z.string(),
   })
   .transform((c) => ({
-    url: c.AUTH_MNGT_SERVICE_URL,
+    url: c.AUTHORIZATION_MANAGEMENT_URL,
   }));
 
 export type AuthServiceConfig = z.infer<typeof AuthServiceConfig>;
